+++ app/domain/show-domain/src/main/java/org/example/entity/genre/Genre.java
@@ -1,4 +1,4 @@
-package org.example.entity;
+package org.example.entity.genre;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
@@ -9,12 +9,14 @@ import jakarta.persistence.Table;
 import lombok.AccessLevel;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
+import org.example.entity.BaseEntity;
+import org.example.entity.show.Show;
 
 @Entity
 @Getter
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
-@Table(name = "app_show_search")
-public class ShowSearch extends BaseEntity {
+@Table(name = "app_genre")
+public class Genre extends BaseEntity {
 
     @Column(name = "name", nullable = false)
     private String name;
