+++ app/domain/show-domain/src/main/java/org/example/entity/show/ShowGenre.java
@@ -1,4 +1,4 @@
-package org.example.entity;
+package org.example.entity.show;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
@@ -7,17 +7,18 @@ import java.util.UUID;
 import lombok.AccessLevel;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
+import org.example.entity.BaseEntity;
 
 @Entity
 @Getter
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
-@Table(name = "app_show_image")
-public class ShowImage extends BaseEntity {
-
-    @Column(name = "url", nullable = false)
-    private String url;
+@Table(name = "app_show_genre")
+public class ShowGenre extends BaseEntity {
 
     @Column(name = "show_id", nullable = false)
     private UUID showId;
+
+    @Column(name = "genre_id", nullable = false)
+    private UUID genreId;
 
 }
